Crianda regra estoque, retornando listas de produtos

+++ InfraBanco/Modelos/TnfEmitente.cs
@@ -24,7 +24,7 @@ namespace InfraBanco.Modelos
 
         [Column("st_ativo")]
         [Required]
-        public short St_Ativo { get; set; }
+        public byte St_Ativo { get; set; }
 
         [Column("apelido")]
         [MaxLength(20)]
